Reduced time between automatic SST put()s to 1ms
The background thread in DerechoGroup that regularly calls put() on the
SST to detect failures was sleeping for 5 seconds at a time. This caused
the DerechoGroup destructor to hang for up to 5s waiting on this thread.
Reducing the timeout makes destruction much faster, and also makes
failure detection more prompt, at the expense of triggering many more
put()s.