Bug fixes to ManagedGroup, adding View serialization
Fixed a long-standing bug in ManagedGroup in which the list of member IP
addresses was never copied from the old view to the new view when
processing a view change. Fixed a bug in the logging system by declaring
the overloaded ostream operator<< inside a namespace. Added
functionality to maintain an additional log file in persistent mode, in
which the current View is serialized whenever a view-change completes.